Pulling updates from Herik

+++ mediajson.lua
@@ -24,6 +24,35 @@ function finishJsonFormGeneral(data)
   return string.gsub("{".. data, '\n"','\n\t"')
 end
 
+function retBoolDot(S,y)
+  y = y * -1
+  return string.sub(S, y, y) == "."
+end
+-- diretorio = arg[1] ...... filename campo da array lista de arquivo ....... i .......
+function genJsonInFolder(directory, filename, lto, limiter)
+  limiter = limiter * -1
+  tempFile = string.sub(filename, 1, limiter)..'.temp'
+  filePath = string.gsub(directory, ' ', '\\ ')..'/'..string.gsub(filename, ' ', '\\ ')
+  os.execute('mediainfo '..filePath..' >> "'..tempFile..'"')
+  isSuccessfull = jsonOutput(jsonProcessing(fileToArray(tempFile), directory, lto), ' '..string.sub(filename, 1, limiter), true)
+  os.execute('rm "'..tempFile..'"')
+  return isSuccessfull
+end
+
+function dirJson(dir, lto)
+  os.execute('ls '..string.gsub(dir, ' ', '\\ ')..' >> templist')
+  filesList = fileToArray("templist")
+  os.execute("rm templist")
+  for i,x in ipairs(filesList) do
+    if string.sub(x, -4, -1) == ".mxf" then
+      isSuccessfull = genJsonInFolder(dir, x, lto, 5)
+    elseif string.sub(x, -3, -1) == ".ts" then
+      isSuccessfull = genJsonInFolder(dir, x, lto, 4)
+    end
+  end
+  return isSuccessfull
+end
+
 --[[ MAIN FUNCTIONS ]]--
 function fileToArray(arquivo, debug)
     if debug == nil then debug = false end
@@ -131,9 +160,16 @@ function jsonProcessing(inputArray, nomeArquivo, lto)
   return finishJsonFormGeneral(outputJsonG)..finishJsonFormatting('Video', outputJsonV)..finishJsonFormatting('Audio', outputJsonA)..finishJsonFormOther(outputJsonO)
 end
 
-function jsonOutput(jsonString)
+function jsonOutput(jsonString, subFile, flagDir)
+  if flagDir then
+    fileName = string.sub(arg[1], 1, -2)..' '..tostring(subFile)..'.json'
+  else
+    fileName = string.sub(arg[1], 1, -5)..'.json'
+  end
 
-  fileName = string.sub(arg[1], 1, -5)..'.json'
+  if not(string.sub(arg[1], -1 , -1) == "/") then
+    fileName = arg[1]..' '..tostring(subFile)..'.json'
+  end
   receiveJson = io.open(fileName, "w")
   receiveJson:write(jsonString)
   receiveJson:close()
@@ -146,10 +182,14 @@ end
 
 assert(not(arg[1] == nil), "Name of the Movie not mentioned in the script arguments! Aborting...")
 assert(not(arg[2] == nil), "Name of the Destination Tape not mentioned in the script arguments! Aborting...")
-tempFile = string.sub(arg[1], 1, -5)..'.temp'
-os.execute('mediainfo "'..arg[1]..'" >> "'..tempFile..'"')
-isSuccessfull = jsonOutput(jsonProcessing(fileToArray(tempFile), arg[1], arg[2]))
-os.execute('rm "'..tempFile..'"')
+if not(retBoolDot(arg[1], 5) or retBoolDot(arg[1], 4) or retBoolDot(arg[1], 3)) then
+  isSuccessfull = dirJson(arg[1], arg[2])
+else
+  tempFile = string.sub(arg[1], 1, -5)..'.temp'
+  os.execute('mediainfo "'..arg[1]..'" >> "'..tempFile..'"')
+  isSuccessfull = jsonOutput(jsonProcessing(fileToArray(tempFile), arg[1], arg[2]), false)
+  os.execute('rm "'..tempFile..'"')
+end
 if isSuccessfull then
     print("JSON creation completed")
     return 1
